+++ app/controllers/replies_controller.rb
@@ -106,13 +106,11 @@ class RepliesController < ApplicationController
 
   def find_editable_reply_or_redirect
     @reply = current_user.editable_replies.find_by(id: params[:id])
-    empty_response_or_root_path if @reply.parent.auto_locked?
     redirect_without_reply
   end
 
   def find_deletable_reply_or_redirect
     @reply = current_user.deletable_replies.find_by(id: params[:id])
-    empty_response_or_root_path if @reply.parent.auto_locked?
     redirect_without_reply
   end